requirements.txt
Django>=4.0,<5.0
mysqlclient>=2.2.0
djangorestframework
djangorestframework-simplejwt>=4.0.0
django-filter
pillow



